Widen instruction/cycle counters to 64-bit, separate types
- Widen g_instructionCount from uint32_t to uint64_t (InstructionCount)
- Introduce ScaledCycleCount (uint64_t) for ICT scheduler cycle math
- Fix truncation bugs in ICTScheduler::add() and doGetNext()
- Fix StopiCount type in m68k_go_nCycles_1
- Update format specifiers to PRIu64 across debugger and core
- Fix test/record.sh ROM path (roms/ -> data/roms/)
- Re-record all 6 golden files
- Add glossary entries: InstructionCount, ScaledCycleCount, kCycleScale

[skip ci]

+++ src/core/extn_extfs.cpp
@@ -19,6 +19,7 @@
 #include "storage/drive_manager.h"
 
 #include <array>
+#include <cinttypes>
 #include <cstdint>
 #include <cstdlib>
 #include <string>
@@ -1070,7 +1071,7 @@ static void RegFatal(uint32_t regParam[], uint16_t &regResult)
 	std::string msg = guestFormatLog(regParam[0], regParam);
 	DIAG(ExtFS, "GUEST FATAL: %s\n", msg.c_str());
 	guestConsoleAppend("FATAL: " + msg);
-	fprintf(stderr, "\n[GUEST FATAL] (insn #%u) %s\n", (unsigned)g_instructionCount, msg.c_str());
+	fprintf(stderr, "\n[GUEST FATAL] (insn #%" PRIu64 ") %s\n", g_instructionCount, msg.c_str());
 	DumpRecentDisasm();
 	fflush(stderr);
 	if (g_debuggerActive)
